fix expo go bundling on Android
